+++ @navikt/core/react/src/form/checkbox/index.ts
@@ -1,5 +1,6 @@
-export { default as Checkbox, type CheckboxProps } from "./Checkbox";
+export { default as Checkbox } from "./Checkbox";
 export {
   default as CheckboxGroup,
   type CheckboxGroupProps,
 } from "./CheckboxGroup";
+export { type CheckboxProps } from "./types";
