+++ @navikt/core/react/src/form/radio/index.ts
@@ -1,2 +1,3 @@
-export { default as Radio, type RadioProps } from "./Radio";
+export { default as Radio } from "./Radio";
 export { default as RadioGroup, type RadioGroupProps } from "./RadioGroup";
+export { type RadioProps } from "./types";
