+++ @navikt/core/react/src/form/search/context.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export interface SearchContextProps {
+  disabled?: boolean;
+  size: "medium" | "small";
+  variant: "primary" | "secondary" | "simple";
+  handleClick: () => void;
+}
+
+export const SearchContext = createContext<SearchContextProps | null>(null);
